backend/requirements.txt
Flask==2.1.1
requests>=2.28.0,<3.0.0
beautifulsoup4==4.10.0
firecrawl-py==1.15.0
python-dotenv==1.1.0
google-genai
